fix: シェアターゲットピッカー不具合修正
